+++ projects/demo/src/app/app.component.ts
@@ -4,6 +4,8 @@ import {
   Component,
   ViewChild,
 } from "@angular/core";
+import { MatChipsModule } from "@angular/material/chips";
+import { MatIconModule } from "@angular/material/icon";
 import { MatPaginator, MatPaginatorModule } from "@angular/material/paginator";
 import { MatTableModule } from "@angular/material/table";
 import {
@@ -11,6 +13,8 @@ import {
   MatMultiSortDirective,
   MatMultiSortHeaderComponent,
   MatMultiSortTableDataSource,
+  MatTableColumnConfigTriggerDirective,
+  TableColumn,
 } from "../../../../src/public-api";
 import { MEMBER_DATA, MemberInformation } from "./data";
 
@@ -22,11 +26,14 @@ const APP_VERSION = "DEBUG";
 @Component({
   selector: "app-root",
   imports: [
+    MatChipsModule,
+    MatIconModule,
     MatTableModule,
     MatPaginatorModule,
     MatMultiSortDirective,
     MatMultiSortControlComponent,
     MatMultiSortHeaderComponent,
+    MatTableColumnConfigTriggerDirective,
   ],
   templateUrl: "./app.component.html",
   styleUrl: "./app.component.scss",
@@ -39,6 +46,34 @@ const APP_VERSION = "DEBUG";
  * @implements {AfterViewInit}
  */
 export class AppComponent implements AfterViewInit {
+  /**
+   * Defines the columns for the table displaying member information.
+   * Each column is represented by an object containing the following properties:
+   * - `id`: A unique identifier for the column.
+   * - `label`: The display name of the column.
+   * - `visible`: A boolean indicating whether the column should be visible.
+   *
+   * The columns are:
+   * - `id`: ID of the member (hidden by default).
+   * - `name`: Name of the member.
+   * - `age`: Age of the member.
+   * - `active`: Indicates if the member is active.
+   * - `joinDate`: The date the member joined.
+   * - `score`: The score of the member.
+   * - `department`: The department the member belongs to.
+   * - `comment`: Additional comments about the member.
+   */
+  readonly columns: TableColumn<MemberInformation>[] = [
+    { id: "id", label: "ID", visible: false },
+    { id: "name", label: "Name", visible: true },
+    { id: "age", label: "Age", visible: true },
+    { id: "active", label: "Active", visible: true },
+    { id: "joinDate", label: "Join Date", visible: true },
+    { id: "score", label: "Score", visible: true },
+    { id: "department", label: "Department", visible: true },
+    { id: "comment", label: "Comment", visible: true },
+  ];
+
   /**
    * The current year based on the system's date.
    * This value is set when the component is initialized.
@@ -62,9 +97,11 @@ export class AppComponent implements AfterViewInit {
    * Each string corresponds to a column identifier.
    *
    */
-  readonly displayedColumns: (keyof MemberInformation)[] = Object.keys(
-    MEMBER_DATA[0]
-  ) as (keyof MemberInformation)[];
+  get displayedColumns(): (keyof MemberInformation)[] {
+    return this.columns
+      .filter((column) => column.visible)
+      .map((column) => column.id);
+  }
 
   /**
    * Reference to the MatPaginator component, used to control pagination in the table.