+++ src/public-api.ts
@@ -2,7 +2,10 @@
  * Public API Surface of lib
  */
 
+export * from "./lib/mat-table-column-config";
 export * from "./lib/mat-multi-sort.directive";
+export * from "./lib/mat-table-column-config-trigger.directive";
 export * from "./lib/mat-multi-sort-table-data-source";
 export * from "./lib/mat-multi-sort-header/mat-multi-sort-header.component";
 export * from "./lib/mat-multi-sort-control/mat-multi-sort-control.component";
+export * from "./lib/mat-table-column-config/mat-table-column-config.component";